Refactored to split out user components into new app pints_user
