+++ pint_journal_project/urls.py
@@ -18,5 +18,6 @@ from django.contrib import admin
 
 urlpatterns = [
     url(r'^admin/', include(admin.site.urls)),
+    url(r'^user/', include('pints_user.urls')),
     url(r'^/?', include('pints_main.urls')),
 ]
