+++ pints_main/forms.py
@@ -1,6 +1,5 @@
 from django import forms
-from pints_main.models import  Brewery, Beer, BeerScore, UserProfile
-from django.contrib.auth.models import User
+from pints_main.models import  Brewery, Beer, BeerScore
 
 class BreweryForm(forms.ModelForm):
 	name = forms.CharField(max_length=128, help_text="Brewery Name")
@@ -40,18 +39,6 @@ class BeerScoreForm(forms.ModelForm):
 	class Meta:
 		model = BeerScore
 		fields = ('score',)
-
-class UserForm(forms.ModelForm):
-	password = forms.CharField(widget=forms.PasswordInput())
-
-	class Meta:
-		model = User
-		fields = ('username', 'email', 'password')
-
-class UserProfileForm(forms.ModelForm):
-	class Meta:
-		model = UserProfile
-		fields = ('picture',)
