+++ pints_main/models.py
@@ -1,7 +1,6 @@
 from django.db import models
 from unidecode import unidecode
 from utils.unique_slugify import unique_slugify
-from django.contrib.auth.models import User
 
 class Brewery(models.Model):
 	name = models.CharField(max_length=128, unique=True)
@@ -70,17 +69,6 @@ class BeerScore(models.Model):
 
 #class Brew_Type_Lookup(models.Mode):
 
-class UserProfile(models.Model):
-	'''
-	Extension of django's built in user model (one-to-one relation)
-	'''
-	user = models.OneToOneField(User)
-
-	# custom user profile fields below
-	picture = models.ImageField(upload_to='profile_images', blank=True)
-
-	def __unicode__(self):
-		return self.user.username
