+++ pints_user/admin.py
@@ -1,3 +1,4 @@
 from django.contrib import admin
+from pints_user.models import UserProfile
 
-# Register your models here.
+admin.site.register(UserProfile)
