+++ pints_user/forms.py
@@ -1,0 +1,20 @@
+from django import forms
+from pints_user.models import UserProfile
+from django.contrib.auth.models import User
+
+class UserForm(forms.ModelForm):
+	password = forms.CharField(widget=forms.PasswordInput())
+
+	class Meta:
+		model = User
+		fields = ('username', 'email', 'password')
+
+class UserProfileForm(forms.ModelForm):
+	class Meta:
+		model = UserProfile
+		fields = ('picture',)
+		
+
+
+
+
