+++ templates/pints_user/register.html
